requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-dbbackup==3.2.0
django-import-export==2.5.0
django-simple-history==3.0.0
django-utils-six==2.0
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
PyMySQL==1.0.2
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
PyYAML==5.4.1
reportlab==3.6.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
webencodings==0.5.1
wfastcgi==3.0.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
autopep8==1.5.7
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
